Started exploring the alternatives
Puredata may be a good choice but it may involve graphical programming
too much to be loved.
SDL seems to be a robust solution to do audio processing in a variety of
languages and on all platforms, with a lot of support on stackoverflow.
Implemented a first example of load and play an audio file in sdl, still
to be completely understood.
Also included an article about libpd.